runtime/vault-signer/pyproject.toml
[project]
name = "vault-signer"
version = "0.1.0"
description = "GCP KMS signing facade for ACHIEVEMOR shield-policy per-substrate attestation"
requires-python = ">=3.12"

dependencies = [
    "fastapi==0.136.0",
    "uvicorn[standard]==0.44.0",
    "google-cloud-kms==3.12.0",
    "google-auth==2.49.2",
    "httpx==0.28.1",
]

[project.optional-dependencies]
dev = [
    "pytest==9.0.3",
    "pytest-asyncio==1.3.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"
